Fixed error with getting a user by their email.
The field in the User object is email. The field in the databasse is
"Email". The getUserByEmail method now works as does the register/login
logic as a result of that. Added postTurn stub and friend info stub.
Uncommented code that was catching all the 404s as that works now as well.