Most fake UI code removed.
Uses all DB Handler methods now. Waiting on a few to be implemented before
this is tested but that is soon. Added getUserByUserName to DBHandler.
Removed all fake/testing/spoof code from Template Handler. Updated
GameView.java to include a boolean to determine if it is the currently
viewing user's turn. Attempted a selector in css for overlapping the cards
in a pile needs testing. Made the createGame template a real thing to
select users and ai players. Un-spoofed friendInfo. Added a form to remove
friend from friends template. Added logic in js in kingscorners.mustache
to allow a move of a pile to another pile. Greyedout cards if not current
turn.