Uses all real DB code can play a complete game.
Edited Card.java to set fancy display stuff (J, Q, K and suit) when
created from DB object. Fixed KCMove to only allow Kings on the Corners
and all numbers on empty piles otherwise. Fixed KingCorner.java to
actually get the currentUser's hand (was null). Handle Move and turn and
render game in TemplateHandler appropriately. Added a CardView to
determine if this is a bottom card for overlaying cards in the template.
Display red cards as red now and properly overlay piles. Hide the not your
turn alert box when not in use.