Fixed redirecting/notification messages.
Removed DBTest call from Main.java. Removed the duplicated TemplateHandler
TJ pushed (it was a really old copy but in a different spot). Removed
break/label statment. Replaced it with boolean flag. Changed
TemplateHandler: for each post method/request set cookies for the
success/error notifications (instead of response headers that went away
after redirect). Now the pre-processing of the ModelAndView also pulls
from the cookies (in addition to the response headers). Changed the 404
erorr handling to include an error. Also displays username by logout
button. Added isActive to GameView. Renamed the static resource folder to
static and the inner folder is public. So static resource dir is static
and stylesheets/js can go under public and not be nuked by 404 exception
handler. The header.mustache had to be changed to include the now
/public/main.css (old was jsut /main.css). The kingscorners.mustache was
edited to disable click handlers for end of game. Reload only if it is not
the viewer's turn (so it wont mess up game play but still auto reload).
Made a better displaying messages system.