Added initial classes for the second game: Circles
Moved classes sepecific to games into their own subfolders.

Changed PileIds to have a function that returns a String of what should be
used for HashMap<String, *> to remove all the
Integer.toString(PileIds.*.ordinal()) calls.

Added CSS for logo and changed the main display name of the site to Grey
Circle Games.

Removed UserManual.pdf that refused to delete.