Issue #19 Give users a username separate from email.
It only looks like I did a bunch because I changed "userName" to "username" and also indentation is gross but not much important stuff here. Added username field to register and login; now username is required and email is not. Also had to fix stuff on adding friends so empty strings don't break it. We may need to change more surrounding this issue but idk whatevs!