Abstracted the Circles board.
Eventually if other similar board games are added, this can be abstracted further - but I figure that not all board games will have the same grid-like layout, so we'd have to add another layer of parent classes for that to work (gross!).
And yes it took me all day to do this. Mongo is rough.