Issue #11 People can now play as a guest.
A guest is just a user but the user naame is Guest ## the email is "" and
the password is Guest ## (shh don't tell anyone - except for the people
that sign in as guests so they can turn their account into theirs).

A "guest" can do everything a normal acount can.