Issue #34 skip player's turn if no moves available.
Currently only relevant to Circles, but can be extended to work in other games. Uses the AI to determine whether there are any valid moves remaining. That will need to change if the AI algorithm changes somehow.
Also updated notifications and emails to let both users know what happened if a turn was skipped. This functionality was not fully tested, so good luck.