Issue #8 Added drag and drop for King's Corner
Need to try this on mobile and possibly add touch events (if it doesn't
introduce yet another js include). Going to try that after pushing this so
I can try on my phone (not on home network can't check on phone
otherwise).

I think for mobile dragging and dropping is not as nice anyway and
clicking is working but for laptops with touchscreens touch events for
drag would be nice.

Also fixed the bug introduceed by Issue #33 and rendering the home page is
appropriately checking if the user is logged in now.