Issue #8 Fixed some bugs with drag
Doesn't freak out whenever you drag an empty pile anymore in fact it is
trying its darndest to not let you drag an empty pile. Fixed a slight css
error in that the background around the dragging item was showing and now
it is not.

There is a nice check in KingsCornerHandler.postMove that makes sure the
moving pile has at least one card - how's that for developer confidence in
the javascript preventing the draging and dropping of emptyPiles? Also now
on drag start it clears out all of the cards that were previously selected
which I think is better in other cases as well.